revert to explicit type kinds (merged Michael's commit by mistake -- too early in the book for that)

+++ src/lib/mod_diff.f90
@@ -1,5 +1,5 @@
 module mod_diff
-use iso_fortran_env, only: wp=>real32
+use iso_fortran_env, only: real32
 
 implicit none
 
@@ -13,11 +13,11 @@ contains
 pure function diffu(x, periodic) result(dx)
   !! Returns a upstream difference of a 1-d array. Applies periodic boundary
   !! conditions if `periodic .eqv. .true.`.
-  real(kind=wp), dimension(:), intent(in) :: x
+  real(kind=real32), dimension(:), intent(in) :: x
     !! Input array
   logical, intent(in), optional :: periodic
     !! If true, periodic boundary conditions area applied. Default false.
-  real(kind=wp), dimension(:), allocatable :: dx
+  real(kind=real32), dimension(:), allocatable :: dx
   logical :: periodic_
   integer :: idm
 
@@ -51,11 +51,11 @@ end function diffu
 pure function diffc(x, periodic) result(dx)
   !! Returns a centered difference of a 1-d array. Applies periodic boundary
   !! conditions if `periodic .eqv. .true.`.
-  real(kind=wp), dimension(:), intent(in) :: x
+  real(kind=real32), dimension(:), intent(in) :: x
     !! Input array
   logical, intent(in), optional :: periodic
     !! If true, periodic boundary conditions area applied. Default false.
-  real(kind=wp), dimension(:), allocatable :: dx
+  real(kind=real32), dimension(:), allocatable :: dx
   logical :: periodic_
   integer :: idm
